refactor(extension): UnauthedScreen を自己完結型に変更し Story を画面ごとに分離
- fetchAuthStatus / AuthStatus 型を authQueries.ts に抽出
- UnauthedScreen が自身でクエリ・ミューテーション・ローカル状態を管理し props を削減
- App.tsx をルーティングのみの薄いオーケストレーターに簡略化
- UnauthedScreen.stories.tsx / MainScreen.stories.tsx を新規作成
- App.stories.tsx をルーティングレベルの2ストーリーに絞り込み
